+++ src/graphql/resolvers/article/getOne.ts
@@ -16,17 +16,41 @@ const getOneArticle: T.Resolver<GraphQL.QueryGetOneArticleArgs, GraphQL.Article 
   const { where } = params;
   const result = await prisma.article.findFirst({
     where,
+    select: {
+      id: true,
+      url: true,
+      ready: true,
+      type: true,
+      isPublished: true,
+      added: true,
+      edited: true,
+      published: true,
+      avgTimeStory: true,
+      avgAllTimeStory: true,
+      created: true,
+      updated: true,
+      ArticleTag: true,
+    },
   });
   if (result === null) {
     return result;
   }
+  // Get article tag list
+  const tags = await prisma.articleTag.findMany({
+    where: {
+      articleId: result.id,
+    },
+    select: {
+      Tag: true,
+    },
+  });
   return {
     id: result.id,
     url: result.url,
-    v: result.v,
+    ready: result.ready || false,
     type: result.type,
-    email: result.email,
-    isPublished: result.isPublished,
+    tags: tags.map((item) => item.Tag.word),
+    isPublished: result.isPublished || false,
     added: result.added.toISOString(),
     edited: result.edited?.toISOString() || '',
     published: result.published?.toISOString() || '',